backend/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
django-filter==24.3
psycopg[binary]==3.2.10
django-environ==0.11.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
celery==5.3.6
redis==5.0.1
drf-yasg==1.21.7
pillow==10.4.0
django-storages==1.14.4
boto3==1.34.150
gunicorn==21.2.0
whitenoise==6.6.0
pytest==8.2.2
pytest-django==4.9.0
factory_boy==3.3.0
requests==2.32.3
responses==0.25.3
django-prometheus==2.3.1
sentry-sdk==1.45.0
python-json-logger==2.0.7
stripe==10.4.0
paypal-checkout-serversdk==1.0.3
setuptools>=65.0.0
